금액 21억이상 처리 변경

+++ src/main/java/com/dksys/biz/user/ar/ar05/service/impl/AR05SvcImpl.java
@@ -66,7 +66,7 @@ public class AR05SvcImpl implements AR05Svc {
 		Map<String, Object> cdtlnData = new HashMap<String, Object>();
 		cdtlnData.putAll(etrdpsData);
 		int clntCd = Integer.parseInt(etrdpsData.get("clntCd"));
-		int etrdpsAmt = Integer.parseInt(etrdpsData.get("etrdpsAmt"));
+		long etrdpsAmt = Long.parseLong(etrdpsData.get("etrdpsAmt"));
 		cdtlnData.put("clntCd", clntCd);
 		cdtlnData.put("etrdpsAmt", etrdpsAmt);
 		ar05Mapper.callCreditLoan(cdtlnData);
